requirements.txt
-e git+https://github.com/olasitarska/django-bootstrap3-datetimepicker#egg=django-bootstrap3-datetimepicker

boto3==1.3.1
click==6.6
cssselect==0.9.2
dj-database-url==0.4.1
django-ckeditor==5.0.3
django-click==1.2.0
django-codemirror-widget
django-countries==3.4.1
django-date-extensions==1.1.0
django-dynamic-fixture==1.9.0
django-gulp-rev==0.2
django-gravatar2==1.4.0
django-leaflet==0.18.1
django-markdown-deux==1.0.5
django-recaptcha
django-sslify>=0.2.0
django-storages-redux==1.3.2
django-suit-redactor==0.0.4
django-suit==0.2.19
Django==1.9.7
djrill==2.1.0
easy_thumbnails==2.3
freezegun
icalendar==3.10
lxml==3.6.0
mandrill==1.0.57
markdown2==2.3.1
model-mommy==1.2.6
mock==2.0.0
opbeat==3.3.4
Pillow==3.3.0
psycopg2==2.6.2
py-trello==0.5.0
pyquery==1.2.13
pytest-cov==2.3.0
pytest-django==2.9.1
pytz==2016.4
requests==2.10
six==1.10.0
slacker==0.9.21
whitenoise==3.2
